# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/authenticate": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authentication - Login Using Mobile",
                "description": "",
                "operationId": "authenticate",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by RFC2616 <https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.4>",
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    },
                    {
                        "name": "App-Version-Name",
                        "in": "header",
                        "description": "Set language parameter by RFC2616 <https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.4>",
                        "schema": {
                            "type": "string",
                            "default": "hayat_delivery_merchant_app"
                        }
                    }
                ],
                "requestBody": {
                    "description": "pass authetication data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "mobile_number",
                                    "firebase_auth_token"
                                ],
                                "properties": {
                                    "country_code": {
                                        "type": "string"
                                    },
                                    "mobile_number": {
                                        "type": "string"
                                    },
                                    "firebase_auth_token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "input validation errors"
                    },
                    "500": {
                        "description": "internal server error"
                    },
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/lookups/getInformation": {
            "get": {
                "tags": [
                    "Lookups"
                ],
                "summary": "Get Information",
                "description": "",
                "operationId": "getInformation",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by RFC2616 <https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.4>",
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "input validation errors"
                    },
                    "500": {
                        "description": "internal server error"
                    },
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/lookups/getBusinessTypes": {
            "get": {
                "tags": [
                    "Lookups"
                ],
                "summary": "Get Business Types",
                "description": "",
                "operationId": "getBusinessTypes",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by RFC2616 <https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.4>",
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "input validation errors"
                    },
                    "500": {
                        "description": "internal server error"
                    },
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
...
        "/api/merchant/storeRating": {
            "post": {
                "tags": [
                    "Merchant"
                ],
                "summary": "Store Rating",
                "description": "",
                "operationId": "storeRating",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by RFC2616 <https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.4>",
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "requestBody": {
                    "description": "pass authetication data in addition to merchant details",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string"
                                    },
                                    "captain_id": {
                                        "type": "integer"
                                    },
                                    "rate": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "input validation errors"
                    },
                    "500": {
                        "description": "internal server error"
                    },
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}